listening-comp/requirements.txt
chromadb
streamlit
youtube-transcript-api
groq
python-dotenv
dataclasses-json
sentence-transformers